+++ frontend/src/components/doctor/DoctorList.js
@@ -1,13 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { getDoctors, deleteDoctor } from '../../services/api';
-import '../../css/PatientForm.css';
 import '../../css/PatientList.css';
 
 const DoctorList = () => {
     const [doctors, setDoctors] = useState([]);
+    const [message, setMessage] = useState('');
+    const navigate = useNavigate();
 
-    useEffect(() => {
+    const fetchDoctors = () => {
         getDoctors().then((response) => {
             if (response.data) {
                 setDoctors(response.data);
@@ -17,39 +18,63 @@ const DoctorList = () => {
         }).catch(error => {
             console.error('There was an error fetching the doctors!', error);
         });
+    };
+
+    useEffect(() => {
+        fetchDoctors();
     }, []);
 
     const handleDelete = (id) => {
-        deleteDoctor(id)
-            .then(() => {
-                setDoctors(doctors.filter(doctor => doctor._id !== id));
-                console.log('Deleted successfully');
-            })
-            .catch(error => console.error('Error deleting:', error));
+        const confirmDelete = window.confirm("Are you sure you want to delete this doctor?");
+        if (confirmDelete) {
+            if (id) {
+                deleteDoctor(id)
+                    .then(response => {
+                        setMessage('Doctor deleted successfully');
+                        setTimeout(() => {
+                            setMessage('');
+                        }, 3000);
+                        fetchDoctors();
+                        navigate('/doctors');
+                    })
+                    .catch(error => {
+                        console.error('Error deleting:', error);
+                        setMessage('Error deleting doctor');
+                    });
+            } else {
+                console.error('ID is undefined');
+            }
+        }
     };
 
     return (
         <center>
             <div className="patient-list-container">
-            <center>
-                <h2>Doctor List</h2>
-                <button type="button">
-                    <Link to="/doctors/new">Add Doctor</Link>
-                </button>
-            </center>
-            {doctors.length > 0 ? (
-                <ul>
-                    {doctors.map((doctor) => (
-                        <li key={doctor._id}>
-                            <Link to={`/doctors/${doctor._id}`}>{doctor.name}</Link>
-                            <button onClick={() => handleDelete(doctor._id)} type="button">Delete</button>
-                        </li>
-                    ))}
-                </ul>
-            ) : (
-                <p>No doctors found</p>
-            )}
-        </div>
+                <h1>Doctors</h1>
+                {message && <p className="message">{message}</p>}
+                <center>
+                    <button type="button" className="add-patient-button">
+                        <Link to="/doctors/new">Add Doctor</Link>
+                    </button>
+                </center>
+                {doctors.length > 0 ? (
+                    <div className="patient-cards">
+                        {doctors.map((doctor) => (
+                            <div key={doctor._id} className="patient-card">
+                                <Link to={`/doctors/${doctor._id}`}>
+                                    <p>{doctor.name}</p>
+                                    <p>Specialty: {doctor.speciality}</p>
+                                </Link>
+                                <button onClick={() => handleDelete(doctor._id)} type="button">
+                                    Delete
+                                </button>
+                            </div>
+                        ))}
+                    </div>
+                ) : (
+                    <p>No doctors found</p>
+                )}
+            </div>
         </center>
     );
 };
