+++ frontend/src/components/hospital/HospitalList.js
@@ -1,13 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { getHospitals, deleteHospital } from '../../services/api';
-import '../../css/PatientForm.css';
 import '../../css/PatientList.css';
 
 const HospitalList = () => {
     const [hospitals, setHospitals] = useState([]);
+    const [message, setMessage] = useState('');
+    const navigate = useNavigate();
 
-    useEffect(() => {
+    const fetchHospitals = () => {
         getHospitals().then((response) => {
             if (response.data) {
                 setHospitals(response.data);
@@ -17,35 +18,63 @@ const HospitalList = () => {
         }).catch(error => {
             console.error('There was an error fetching the hospitals!', error);
         });
+    };
+
+    useEffect(() => {
+        fetchHospitals();
     }, []);
 
     const handleDelete = (id) => {
-        deleteHospital(id)
-            .then(() => {
-                setHospitals(hospitals.filter(hospital => hospital._id !== id));
-                console.log('Deleted successfully');
-            })
-            .catch(error => console.error('Error deleting:', error));
+        const confirmDelete = window.confirm("Are you sure you want to delete this hospital?");
+        if (confirmDelete) {
+            if (id) {
+                deleteHospital(id)
+                    .then(response => {
+                        setMessage('Hospital deleted successfully');
+                        setTimeout(() => {
+                            setMessage('');
+                        }, 3000);
+                        fetchHospitals();
+                        navigate('/hospitals');
+                    })
+                    .catch(error => {
+                        console.error('Error deleting:', error);
+                        setMessage('Error deleting hospital');
+                    });
+            } else {
+                console.error('ID is undefined');
+            }
+        }
     };
 
     return (
         <center>
             <div className="patient-list-container">
-            <h2>Hospitals</h2>
-            <button type="button"><Link to="/hospitals/new">Add Hospital</Link></button>
-            {hospitals.length > 0 ? (
-                <ul>
-                    {hospitals.map((hospital) => (
-                        <li key={hospital._id}>
-                            <Link to={`/hospitals/${hospital._id}`}>{hospital.name}</Link>
-                            <button onClick={() => handleDelete(hospital._id)}>Delete</button>
-                        </li>
-                    ))}
-                </ul>
-            ) : (
-                <p>No hospitals found</p>
-            )}
-        </div>
+                <h1>Hospitals</h1>
+                {message && <p className="message">{message}</p>}
+                <center>
+                    <button type="button" className="add-patient-button">
+                        <Link to="/hospitals/new">Add Hospital</Link>
+                    </button>
+                </center>
+                {hospitals.length > 0 ? (
+                    <div className="patient-cards">
+                        {hospitals.map((hospital) => (
+                            <div key={hospital._id} className="patient-card">
+                                <Link to={`/hospitals/${hospital._id}`}>
+                                    <p>{hospital.name}</p>
+                                    <p>Address: {hospital.address}</p>
+                                </Link>
+                                <button onClick={() => handleDelete(hospital._id)} type="button">
+                                    Delete
+                                </button>
+                            </div>
+                        ))}
+                    </div>
+                ) : (
+                    <p>No hospitals found</p>
+                )}
+            </div>
         </center>
     );
 };
